+++ packages/protocol/src/index.ts
@@ -1,7 +1,0 @@
-export * from './version';
-export * from './limits';
-export * from './commonSchemas';
-export * from './resourceSchemas';
-export * from './publicEventSchemas';
-export * from './internalWorkerSchemas';
-export * from './workTypes';